backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
boto3==1.35.19
botocore==1.35.19
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cohere==5.9.2
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.2.2
fastavro==1.9.7
filelock==3.16.0
Flask==3.0.3
fonttools==4.53.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.24.7
idna==3.9
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
parameterized==0.9.0
pillow==10.4.0
pydantic==2.9.1
pydantic_core==2.23.3
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
threadpoolctl==3.5.0
tokenizers==0.20.0
tqdm==4.66.5
types-requests==2.32.0.20240914
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
